+++ components/ExampleModule/index.js
