+++ pages/index.js
@@ -1,7 +1,6 @@
 import { gql } from "@apollo/client";
 import client from "../app/apollo-client";
-import { useEffect } from "react";
-
+import { WhatsNext } from "@/components/ExampleModule";
 export async function getStaticProps() {
   const { data } = await client.query({
     query: gql`
@@ -31,26 +30,10 @@ export async function getStaticProps() {
 }
 
 export default function Home({ posts }) {
-  useEffect(() => {
-    console.log(posts);
-  }, [posts]);
   return (
     <>
-      <h1>Next.js + Apollo Client</h1>
-      <div className="posts">
-        {posts.map((post) => (
-          <div key={post.id} className="post">
-            <h2>{post.title}</h2>
-
-            <h3>{post.author.node.name}</h3>
-
-            <div dangerouslySetInnerHTML={{ __html: post.excerpt }} />
-          </div>
-
-          
-        ))}
-      </div>
-      
+      <h1>Next.js + Apollo Client</h1>    
+      <WhatsNext />  
     </>
     
   );
